Output vector is the result produced when you run a function or perform an operation in your R script, it is the dtaa that R gives back to you after you've given it some input or instructions
Firstly i had created an empty vector called 'output'. I create an input vector and put in the following colours. I used tyhe loop function of for() to loop the values inside of input, in order to count the characters in each colour

In the next exercise I had created a vector with words i want to add an additional value to. In this case I will add the word 'word' to the end of each of the values. Paste() function is used to bring together strings to create a single character vector.
1. Put the strings you want to concatenate
2. Sep = specifies the separator to be used between the strings

Conditional statements are used to pick out certain values if they are true.