The while loop should be used for code within curly braclets and they will continue to repeat while the statement is TRUE. e.g. while(x is less than 5) the loop will continue.
x %% 5 != 0 this expression checks if the remainder when x is divided by 5 is not zero. If it is zero, then x is a factor of 5, otherwise it isn't

Functions can be used instead of writing the same code each time/ copying and pasting. In the case of powers, we have made a function called powers that takes one argument of "x". return (y) will return the value of y, which is the square of x

month_converter is the name of the function and it takes one argument, known as "month" and this represents the month name as a string.
- grep function searches for the "month" string within the "months" vector. it returns the position (indices) where the string is found
- pattern = month specifies the target string to search for month
- x = months defines the vector where the search occurs (months)

numerical_position <- month_converter(month) calls the month_converter function with the specified month and stores the returned value in numerical_position.